Added comment function

+++ personal-blog/src/posts/views.py
@@ -1,7 +1,7 @@
 from urllib import quote_plus
 
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.shortcuts import render
@@ -10,6 +10,11 @@ from .forms import PostForm
 from .models import Post
 
 def post_create(request):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	
+	#if not request.user.is_authenticated():
+		#raise Http404
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -56,6 +61,8 @@ def post_list(request): #list items
 
 
 def post_update(request, slug=None):
+	if request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -72,6 +79,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+	if request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully Created")
